// client/src/validations/authValidation.ts
import { z } from "zod";

export const registerSchema = z.object({
    name: z.string().min(2, "Name is required"),

    email: z.string().email("Invalid email address"),

    password: z.string().min(6, "Password must be at least 6 characters"),
});

export const loginSchema = z.object({
    email: z.string().email("Invalid email"),

    password: z.string().min(6, "Password is required"),
});